Use pkg-config & add a check for MacPorts
Using `pkg-config` will be the most flexible option, and will support a wider range of users with non-standard configurations (not just Homebrew & MacOS). This *should* also work for Homebrew and MacPorts users.

We can fall back to searching for GMP using the Homebrew and MacPorts paths just in case `pkg-config` isn't installed.